requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.10.11
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.5.2
async-timeout==4.0.3
attrs==25.3.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.1.8
dataclasses-json==0.6.7
distro==1.9.0
dnspython==2.6.1
exceptiongroup==1.3.0
fastapi==0.110.2
filelock==3.16.1
frozenlist==1.5.0
fsspec==2025.3.0
greenlet==3.1.1
grpcio==1.70.0
grpcio-tools==1.70.0
h11==0.16.0
h2==4.1.0
hf-xet==1.1.8
hpack==4.0.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.34.4
hyperframe==6.0.1
idna==3.10
Jinja2==3.1.3
jiter==0.9.1
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.2.17
langchain-community==0.2.19
langchain-core==0.2.43
langchain-huggingface==0.0.3
langchain-text-splitters==0.2.4
langsmith==0.1.147
MarkupSafe==2.1.5
marshmallow==3.22.0
mpmath==1.3.0
multidict==6.1.0
mypy_extensions==1.1.0
networkx==3.1
numpy==1.24.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.9.86
nvidia-nvtx-cu12==12.1.105
openai==1.99.9
orjson==3.10.15
packaging==23.2
pika==1.3.2
pillow==10.4.0
pkg_resources==0.0.0
propcache==0.2.0
protobuf==5.29.5
pydantic==1.10.22
pydantic_core==2.27.2
pymongo==4.6.3
PyMySQL==1.1.0
python-dotenv==1.0.1
PyYAML==6.0.2
qdrant-client==1.1.0
regex==2024.11.6
requests==2.31.0
requests-toolbelt==1.0.0
safetensors==0.5.3
scikit-learn==1.3.2
scipy==1.10.1
sentence-transformers==3.2.1
sniffio==1.3.1
SQLAlchemy==2.0.30
starlette==0.37.2
sympy==1.13.3
tenacity==8.5.0
threadpoolctl==3.5.0
tokenizers==0.20.3
torch==2.3.1
tqdm==4.67.1
transformers==4.46.3
triton==2.3.1
typing-inspect==0.9.0
typing_extensions==4.13.2
urllib3==1.26.20
uvicorn==0.29.0
yarl==1.15.2
